+++ src/embed/embed.go
@@ -1,0 +1,6 @@
+package embed
+
+import "embed"
+
+//go:embed *
+var Dir embed.FS