+++ src/internal/app/transports/rest_api/config.go
@@ -1,4 +1,0 @@
-package rest_api
-
-type Config struct {
-}