+++ src/internal/app/transports/rest_api/engine.go
@@ -2,20 +2,60 @@ package rest_api
 
 import (
 	"context"
-	"github.com/gofiber/fiber/v3"
+	"sm-box/src/internal/app/transports/rest_api/config"
+	"sm-box/src/pkg/core/components/logger"
+	"sm-box/src/pkg/core/components/tracer"
+)
+
+const (
+	loggerInitiator = "transports-[http]=rest_api"
 )
 
 type Engine interface {
-	Serve() (err error)
+	Listen() (err error)
 	Shutdown() (err error)
 }
 
-func New(ctx context.Context) (eng Engine, err error) {
-	var e = new(engine)
+func New(ctx context.Context, conf *config.Config) (eng Engine, err error) {
+	// tracer
+	{
+		var trc = tracer.New(tracer.LevelMain, tracer.LevelTransport)
 
-	e.app = fiber.New()
+		trc.FunctionCall(ctx, conf)
+		trc.Error(err).FunctionCallFinished(eng)
+	}
 
-	eng = e
+	var ref = &engine{
+		conf: conf,
+		ctx:  ctx,
+	}
+
+	ref.conf.FillEmptyFields()
+
+	// Компоненты
+	{
+		ref.components = new(components)
+
+		// Logger
+		{
+			if ref.components.logger, err = logger.New(loggerInitiator); err != nil {
+				return
+			}
+		}
+	}
+
+	// fiber
+	{
+		if err = ref.initFiberApp(); err != nil {
+			return
+		}
+	}
+
+	ref.components.logger.Info().
+		Text("The http rest engine has been created. ").
+		Field("config", ref.conf).Write()
+
+	eng = ref
 
 	return
 }