+++ src/pkg/core/states.go
@@ -8,6 +8,14 @@ const (
 	StateOff
 )
 
+var statesList = [...]string{
+	"Nil",
+	"New",
+	"Booted",
+	"Served",
+	"Off",
+}
+
 // State - состояние ядра системы.
 // Возможные варианты:
 //  0. StateNil    - "Nil";
@@ -18,18 +26,9 @@ const (
 type State int
 
 // String - получение строкового представления состояния ядра системы.
-func (index State) String() (v string) {
-	switch index {
-	case StateNil:
-		v = "Nil"
-	case StateNew:
-		v = "New"
-	case StateBooted:
-		v = "Booted"
-	case StateServed:
-		v = "Served"
-	case StateOff:
-		v = "Off"
+func (i State) String() (v string) {
+	if i >= StateNil && int(i) < len(statesList) {
+		return statesList[i]
 	}
 
 	return