+++ src/pkg/tools/http/charsets.go
@@ -1,0 +1,5 @@
+package http_tools
+
+const (
+	CharsetUTF8 = "utf-8"
+)