+++ src/pkg/tools/http/mime.go
@@ -1,0 +1,19 @@
+package http_tools
+
+const (
+	MIMETextXML        = "text/xml"
+	MIMETextJSON       = "text/json"
+	MIMETextYAML       = "text/yaml"
+	MIMETextHTML       = "text/html"
+	MIMETextPlain      = "text/plain"
+	MIMETextJavaScript = "text/javascript"
+
+	MIMEApplicationXML  = "application/xml"
+	MIMEApplicationJSON = "application/json"
+	MIMEApplicationYAML = "application/yaml"
+
+	MIMEApplicationJavaScript = "application/javascript"
+	MIMEApplicationForm       = "application/x-www-form-urlencoded"
+	MIMEOctetStream           = "application/octet-stream"
+	MIMEMultipartForm         = "multipart/form-data"
+)